test: Reorganize unit tests
